add in jgit test code.
Adds in the start of a jgit based history rewrite class.

Right now it just does a git log on the current branch.

The next step is to create a new chain of commits with the same tree's but with new commit messages.

If that works then we will iterate overall local branches and rewrite the history.

If it doesn't work we can look at using git filter-branch --env-filter and a generated filter script for the usersnames we need to rewrite.