use jgit to rewrite history
This version will create a new branch at the first revision of master.

Then it will replay all of the commits and commit new commits with the same
tre e but with altered author and committer fields.

When testing this don't point at this repo but at another read only clone.