add initial support for branch rewrite and extract
Adjusts slightly the branch rewrite logic to fit with the new structure.

The idea is that the whole repo is checked out into a single master branch from svn.  Then the rewrite functionality is used to creeate a new branch with the author corrected and comment corrected details.

The next step is to use the extract functionality to slice off a part of the tree and turn it into a new branch.

Both of these parts are still a work in progress but I need to test in Linux since windows 7 performance on the 3GB wicketstuff repository is mind numbing.